# openApi.yaml
openapi: 3.1.0
info:
  title: Budget Tracking API
  description: API for managing income, expenses, recurring and one-time budgets, and financial summaries.
  version: 1.5.0
  contact:
    name: John Doe
    email: "test@test.com"
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /register:
    post:
      description: Register a new user
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUser"
      responses:
        "201":
          description: User registered successfully
        "400":
          description: Invalid input
  /login:
    post:
      description: Authenticate user and return JWT token
      operationId: loginUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUser"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized - Invalid credentials
  /logout:
    post:
      description: Logout the current user
      operationId: logoutUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully logged out"
        "401":
          description: Unauthorized - Invalid or missing token
        "400":
          description: Bad Request - Token validation failed
  /income:
    post:
      description: Add a new income record
      operationId: addIncomeRecord
      tags:
        - Income
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeRequest"
      responses:
        "201":
          description: Income record created successfully
        "400":
          description: Invalid input
    get:
      description: Get all income records
      operationId: getAllIncomeRecords
      tags:
        - Income
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of income records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeRecordResponse"
  /income/{id}:
    get:
      description: Get an income record by ID
      operationId: getIncomeByID
      tags:
        - Income
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "200":
          description: Income record details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeRecordResponse"
        "404":
          description: Income record not found
    put:
      description: Update an existing income record
      operationId: updateIncomeRecord
      tags:
        - Income
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeRequest"
      responses:
        "200":
          description: Income record updated successfully
        "400":
          description: Invalid input
    delete:
      description: Delete an income record
      operationId: deleteIncomeRecord
      tags:
        - Income
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Income record deleted successfully
  /expenses:
    post:
      description: Add a new expense record
      operationId: addExpenseRecord
      tags:
        - Expenses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseRecord"
      responses:
        "201":
          description: Expense record created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecordResponse"
    get:
      description: Get all expense records
      operationId: getAllExpenseRecords
      tags:
        - Expenses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of expense records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseRecordResponse"
  /expenses/{id}:
    get:
      description: Get an expense record by ID
      operationId: getExpenseByID
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "200":
          description: Expense record found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecordResponse"
        "404":
          description: Expense record not found
    put:
      description: Update an expense record
      operationId: updateExpense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseRecord"
      responses:
        "200":
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecordResponse"
    delete:
      description: Delete an expense record
      operationId: deleteExpense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "204":
          description: Expense deleted successfully
  /expenses/category/{category}:
    get:
      description: Get expenses by category
      operationId: getExpensesByCategory
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "200":
          description: List of expenses for the category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseRecordResponse"
  /expenses/range:
    get:
      description: Get expenses within a date range
      operationId: getExpensesByDateRange
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: 2024-01-01T00:00:00Z
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: 2024-01-31T23:59:59Z
      responses:
        "200":
          description: List of expenses within date range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseRecordResponse"
  /expenses/recent:
    get:
      description: Get recent expenses
      operationId: getRecentExpenses
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: List of recent expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseRecordResponse"
  /expenses/monthly/total:
    get:
      description: Get monthly expense total
      operationId: getMonthlyExpenseTotal
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: date-time
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: 2024-01-01T00:00:00Z
      responses:
        "200":
          description: Monthly expense total
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_amount:
                    type: number
                    format: float
                  currency:
                    type: string
  /expenses/category/totals:
    get:
      description: Get expense totals by category
      operationId: getExpenseTotalsByCategory
      tags:
        - Expenses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Expense totals by category
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    category_id:
                      type: string
                      format: uuid
                    category_name:
                      type: string
                    currency:
                      type: string
                    transaction_count:
                      type: integer
                    total_amount:
                      type: number
                      format: float
  /budgets:
    post:
      description: Create a new budget
      operationId: addBudgetRecord
      tags:
        - Budgets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetRecord"
      responses:
        "201":
          description: Budget created successfully
        "400":
          description: Invalid input
    get:
      description: List all budgets for the user
      operationId: listBudgets
      tags:
        - Budgets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of all budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetRecordResponse"
        "401":
          description: Unauthorized
  /budgets/{id}:
    get:
      description: Get budget usage details by ID
      operationId: getBudgetUsage
      tags:
        - Budgets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "200":
          description: Budget usage details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetUsageResponse"
        "400":
          description: Invalid budget ID
        "401":
          description: Unauthorized
        "404":
          description: Budget not found
    put:
      description: Update an existing budget
      operationId: updateBudget
      tags:
        - Budgets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetRecord"
      responses:
        "200":
          description: Budget updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetRecordResponse"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Budget not found
    delete:
      description: Delete a budget
      operationId: deleteBudget
      tags:
        - Budgets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "204":
          description: Budget deleted successfully
        "400":
          description: Invalid budget ID
        "401":
          description: Unauthorized
        "404":
          description: Budget not found
  /budgets/recurring:
    get:
      description: Get all recurring budgets for the user
      operationId: getRecurringBudgets
      tags:
        - Budgets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of recurring budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetRecordResponse"
        "401":
          description: Unauthorized
  /budgets/one-time:
    get:
      description: Get all recurring budgets for the user
      operationId: getOneTimeBudgets
      tags:
        - Budgets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of recurring budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetRecordResponse"
        "401":
          description: Unauthorized
  /budgets/category/{categoryId}:
    get:
      description: Get budgets for a specific category
      operationId: getBudgetsByCategory
      tags:
        - Budgets
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of budgets for the category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetRecordResponse"
        "400":
          description: Invalid category ID
        "401":
          description: Unauthorized
  /budgets/alerts:
    get:
      description: Get budgets that are near or over their limit
      operationId: getBudgetsNearLimit
      tags:
        - Budgets
      security:
        - bearerAuth: []
      parameters:
        - name: alert_threshold
          in: query
          required: false
          schema:
            type: number
            minimum: 0
            default: 80
      responses:
        "200":
          description: List of budgets near their limit
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetAlertResponse"
        "401":
          description: Unauthorized
  /summary/monthly:
    get:
      description: Get a monthly financial summary
      operationId: getMonthlySummary
      tags:
        - Summary
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: 2024-01-01T00:00:00Z
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Monthly summary of income, expenses, and budgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySummary"
  /summary/yearly:
    get:
      description: Get a yearly financial summary
      operationId: getYearlySummary
      tags:
        - Summary
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
            example: 2024
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Yearly summary of income, expenses, and budgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearlySummary"
...
  /user/profile:
    get:
      description: Get user profile information
      operationId: getUserProfile
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
    put:
      description: Update user profile information
      operationId: updateUserProfile
      tags:
        - User
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
        "400":
          description: Invalid input
  /user/password:
    put:
      description: Update user password
      operationId: updateUserPassword
      tags:
        - User
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "200":
          description: Password updated successfully
        "401":
          description: Unauthorized or invalid current password
        "400":
          description: Invalid input
  /user/stats:
    get:
      description: Get user statistics
      operationId: getUserStats
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "401":
          description: Unauthorized
  /user/roles:
    get:
      description: Get user role information
      operationId: getUserRole
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User role retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleResponse"
        "401":
          description: Unauthorized
        "404":
          description: User role not found
    put:
      description: Update user role
      operationId: updateUserRole
      tags:
        - User
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
      responses:
        "200":
          description: User role updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleResponse"
        "401":
          description: Unauthorized
        "403":
          description: Permission denied - Admin only operation
        "404":
          description: User or role not found
  /user/roles/list:
    get:
      description: List users by role
      operationId: listUsersByRole
      tags:
        - User
      security:
        - bearerAuth: []
      parameters:
        - name: role_name
          in: query
          required: true
          schema:
            type: string
            enum: [Admin, User, Guest, Moderator, Editor, Viewer, Manager]
            example: Admin
      responses:
        "200":
          description: List of users with specified role
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRoleListItem"
        "401":
          description: Unauthorized
        "404":
          description: No users found for specified role
  /categories:
    post:
      description: Create a new category
      operationId: createCategory
      tags:
        - Categories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          description: Invalid input
        "409":
          description: Category already exists
    get:
      description: List all categories for the user
      operationId: listCategories
      tags:
        - Categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
  /categories/{id}:
    get:
      description: Get a category by ID
      operationId: getCategoryById
      tags:
        - Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          description: Category not found
    put:
      description: Update a category
      operationId: updateCategory
      tags:
        - Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          description: Category not found
    delete:
      description: Delete a category
      operationId: deleteCategory
      tags:
        - Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Category deleted successfully
        "404":
          description: Category not found
  /categories/{id}/stats:
    get:
      description: Get usage statistics for a category
      operationId: getCategoryStats
      tags:
        - Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3c94bf9c-15a9-4137-a317-9d78db0a5dbd"
      responses:
        "200":
          description: Category usage statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryStats"
        "404":
          description: Category not found
  /categories/stats/most-used:
    get:
      description: Get most used categories
      operationId: getMostUsedCategories
      tags:
        - Categories
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 100
          description: Maximum number of categories to return
      responses:
        "200":
          description: List of most used categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryUsageStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterUser:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: password
      required:
        - name
        - email
        - password
    LoginUser:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: password
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    CreateIncomeRequest:
      type: object
      properties:
        amount:
          type: number
          format: float
          example: 1500.00
        currency:
          type: string
          example: USD
        source:
          type: string
          example: Salary
        date:
          type: string
          format: date-time
          example: 2024-01-15T00:00:00Z
        description:
          type: string
          example: Monthly salary for January
      required:
        - amount
        - currency
        - source
        - date
    IncomeRecordResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          format: float
          example: 1500.00
        currency:
          type: string
          example: USD
        source:
          type: string
          example: Salary
        date:
          type: string
          format: date-time
          example: 2024-01-15T00:00:00Z
        description:
          type: string
          example: Monthly salary for January
    ExpenseRecord:
      type: object
      properties:
        amount:
          type: number
          format: float
          example: 100.00
        currency:
          type: string
          example: USD
        category_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        date:
          type: string
          format: date-time
          example: 2024-01-16T00:00:00Z
        description:
          type: string
          example: Grocery shopping at local market
      required:
        - amount
        - currency
        - category
        - date
    ExpenseRecordResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          format: float
          example: 100.00
        currency:
          type: string
          example: USD
        category_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        date:
          type: string
          format: date-time
          example: 2024-01-16T00:00:00Z
        description:
          type: string
          example: Grocery shopping at local market
    BudgetRecord:
      type: object
      properties:
        amount:
          type: number
          format: float
          example: 500.00
        currency:
          type: string
          example: USD
        category_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        type:
          type: string
          enum: [recurring, one-time]
          example: recurring
        start_date:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        end_date:
          type: string
          format: date-time
          example: 2024-12-31T00:00:00Z
      required:
        - amount
        - currency
        - category
        - type
        - start_date
...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
...